refactor(analyze): split service/mod.rs into events/routes/tests submodules (line-cap)
service/mod.rs (2163 lines, at its allowlist budget) is split into eleven
focused files, all under the 500-line cap:

  service/
    events.rs          287  AnalyzerEvent, AnalyzerAppState, ApiError, fetch_chunks
    routes.rs          209  build_router, serve, health, list_indexes, sse_handler
    handlers/
      mod.rs            20  submodule declarations
      analysis.rs      296  complexity, smells, quality, refactor, diagnostics
      graph.rs         337  KG graph, entities, clusters, NER, SCIP ingest
      facts.rs         122  facts CRUD
      review.rs        255  diff review, GitHub PR review, webhooks
      deep.rs          272  LLM deep analysis, synthesise_review_from_chunks, lookup_frameworks
    tests.rs           271  health/SSE/facts/SCIP/diagnostics/index-proxy tests
    tests_review.rs    318  review/deep/webhook tests
    mod.rs              52  thin re-export facade

Removes the allowlist entry for service/mod.rs from .line-cap-allowlist.tsv
(170 entries remain; ratchet ticks down by 1).

All 357 tests pass. Zero clippy warnings. No behaviour change.

Closes #958
