+++ .line-cap-allowlist.tsv
@@ -24,7 +24,6 @@ crates/trusty-analyze/src/lang/adapters/typescript.rs	680
 crates/trusty-analyze/src/lang/detection.rs	616
 crates/trusty-analyze/src/main.rs	849
 crates/trusty-analyze/src/mcp/mod.rs	1178
-crates/trusty-analyze/src/service/mod.rs	2163
 crates/trusty-analyze/tests/stdio_harness.rs	639
 crates/trusty-common/src/bm25_client.rs	503
 crates/trusty-common/src/bm25.rs	645
@@ -140,7 +139,7 @@ crates/trusty-search/src/commands/doctor_checks.rs	612
 crates/trusty-search/src/commands/integrate.rs	688
 crates/trusty-search/src/commands/reindex_engine.rs	1862
 crates/trusty-search/src/commands/reindex_ui.rs	1458
-crates/trusty-search/src/commands/start.rs	2197
+crates/trusty-search/src/commands/start.rs	1892
 crates/trusty-search/src/core/chunker/mod.rs	1883
 crates/trusty-search/src/core/classifier.rs	1032
 crates/trusty-search/src/core/corpus.rs	1685
